Added some support for the vt_tileset patch for NetHack.
The tileset is currently added to the swf as a resource (and thus can't
be changed by the user). It should be downloaded from the server instead
to avoid bloating the binary for users that don't want it.
The tile size is also hardcoded.

I've only done limited testing of it.



git-svn-id: svn://localhost/soiled@100 c1b1574c-8591-4b2f-8170-dace01c35af4